Day 11 Part 2 Solved
Not Solved:
Actually predicted this one coming but didn't plan for it. Also trying something new because of out of memory error.

Day 11 Part 2 Solved:

The refactor worked. Who needs to store the actual image, not I, just the galaxy locations. Super easy to trim. ALSO got it first try with real data!!