Day 12 Part 1 Solved
That took a bit. Some distraction with watching survivor. Some more time by trying be fancy with regex and that took too much time and failed. Happy with using binary to get all possible combos. Kinda/maybe brute force but worked. Also may be able to reduce the bruteness of the force lol