Day 12 Part 2 SOLVED!
Finally solved this one. had to use some reddit and finally learning how to use the Visual Studio debugger which show that I need to change ONLY ONE line of code to get it working!!!!! Needed to make the list in the method new and not a ref to another one for the recursion to work.