21 Day 6 Part 2 Complete
FIRST TRY!!!

Part 1 way of doing this would run out of memory on the TEST input!!!

Had to refactor how I was doing it. Figured this out without ANY help from reddit or online in general. Very happy with myself.

End result only took 00.0004995 to complete for a 1.7 trillion number