24 Day 3 Part 2 Complete
Not first try :(

First problem that took way too long was the fact that there was a different testInput for part 2 then part 1. This cost sooo much time, where I thought i was not understanding regex again but in fact it can't find "do()" if there is not "do()" in the testInput lol.

Then I messed up on a simple mistake of making either do() or don't() flip the bool. Whoops.