24 Day 5 Part 2 Complete
First Try!

Was not too bad but forgot when making copy of list to make sure it's not a ref to the old.

This probably should been done fast but I had the main concept mostly done it was writing it, refactoring the old code to use lists, and debugging some simple mistakes led to extra time to complete