24 Day 6 Part 1 Complete
Not First Try :(

Problem with using properties of the class outside of the Part1 and Part2 methods. If I don't "reset" them then the test example might mess with the numbers when I run the real. For this the count of the test was added to the real walk because I didn't reset the counter to 0. Whoops. Simple fix.