24 Day 6 Part 2 COMPLETE
FINALLY Complete with try of 1793.

Used some reddit comments to get me on the right track. I was close but my logic was messed up slightly.

When I turned on the existing blocks for the original path I wasn't properly logging that dir. So when going through, it was not checking the spot just after the turn as a fake obj. WHOOPS.