24 Day 7 Part 1 Complete
Not First try but it was simple fix.

I needed to upgrade the answer int to a long but when I did I got to adjust the testAnswer in the function to long too so first attempt was doing weird out of bounds math with int number not going high enough.