24 Day 9 Part 2 Complete
First Try!

Interesting note: this took unity ~4.5secs when using the debugger mode but only <1 sec when not using the debugger mode.

Took a bit of debugging to get it to work but all the problems where simple mistakes on my end of coding but the overall logic and idea was right and unchanged.

I did have to basically refactor most of part 1 to do part 2 so a lot more time.