24 Day 10 Part 1 Complete
First try...

Could have done this HOURS ago. But I tried to had cache of the different nodes in the trail to make future searches faster and avoid redoing prev checks. However I couldn't figure it out. I think it was me using the trailNodes as a ref was screwing me up but my brain was too fried to un mess it up.