24 Day 11 Part 2 Refactor
I refactor to trying breaking apart the proccess when number of stone gets too large.

This is hopefully to help memory sizes because I was running out.

Note this new way worked with Part 1 testExample and realInput in both cases the new refactor took longer but Idk about memory yet